Select Location component refactored
