+++ GoToProject/src/library/hooks/index.js
@@ -1,0 +1,1 @@
+export {useMap} from './useMap.js';
